Refactor model structure
- Move models into subdirectories
- Split models into one file per class
