+++ backend/cms/forms/extra_templates/extra_template_form.py
@@ -3,7 +3,7 @@ Form for creating an extra_template object
 """
 from django import forms
 
-from ...models.extra_template import ExtraTemplate
+from ...models import ExtraTemplate
 from ...utils.slug_utils import generate_unique_slug
