+++ backend/cms/forms/media/__init__.py
@@ -1,0 +1,1 @@
+from .document_form import DocumentForm
