+++ backend/cms/forms/media/document_form.py
@@ -1,0 +1,10 @@
+from django import forms
+
+from ...models import Document
+
+
+class DocumentForm(forms.ModelForm):
+
+    class Meta:
+        model = Document
+        fields = ('description', 'document', )
