+++ backend/cms/forms/pages/__init__.py
@@ -1,1 +1,2 @@
-from .page_form import PageForm, PageTranslationForm
+from .page_form import PageForm
+from .page_translation_form import PageTranslationForm
