+++ backend/cms/forms/pois/__init__.py
@@ -1,1 +1,2 @@
-from .poi_form import POIForm, POITranslationForm
+from .poi_form import POIForm
+from .poi_translation_form import POITranslationForm
