+++ backend/cms/forms/users/__init__.py
@@ -1,2 +1,4 @@
-from .user_form import UserForm, UserProfileForm
-from .region_user_form import RegionUserForm, RegionUserProfileForm
+from .user_form import UserForm
+from .user_profile_form import UserProfileForm
+from .region_user_form import RegionUserForm
+from .region_user_profile_form import RegionUserProfileForm
