+++ backend/cms/locale/de/LC_MESSAGES/django.po
@@ -7,7 +7,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: 1.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-12-22 18:36+0000\n"
+"POT-Creation-Date: 2019-12-22 22:45+0000\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: Timo Ludwig <ludwig@integreat-app.de>\n"
 "Language-Team:\n"
@@ -188,7 +188,7 @@ msgstr "Entwurf"
 msgid "Pending Review"
 msgstr "Review ausstehend"
 
-#: constants/status.py:11 forms/pois/poi_form.py:71
+#: constants/status.py:11 forms/pois/poi_translation_form.py:25
 msgid "Public"
 msgstr "Öffentlich"
 
@@ -306,7 +306,7 @@ msgstr "Gespiegelte Seite vor dieser Seite einbinden"
 msgid "Embed mirrored page after this page"
 msgstr "Gespiegelte Seite nach dieser Seite einbinden"
 
-#: forms/pois/poi_form.py:72
+#: forms/pois/poi_translation_form.py:26
 msgid "Private"
 msgstr "Privat"
 
@@ -314,11 +314,11 @@ msgstr "Privat"
 msgid "Do no import initial content"
 msgstr "Keine Inhalte übernehmen"
 
-#: models/language.py:24
+#: models/languages/language.py:19
 msgid "Left to right"
 msgstr "Links nach Rechts"
 
-#: models/language.py:25
+#: models/languages/language.py:20
 msgid "Right to left"
 msgstr "Rechts nach Links"
