+++ backend/cms/models/__init__.py
@@ -1,28 +1,36 @@
-from .configuration import Configuration
-from .event import Event
-from .event import EventTranslation
-from .event import RecurrenceRule
-from .extra import Extra
-from .extra_template import ExtraTemplate
-from .feedback import EventFeedback
-from .feedback import ExtraFeedback
-from .feedback import Feedback
-from .feedback import PageFeedback
-from .feedback import RegionFeedback
-from .feedback import SearchResultFeedback
-from .feedback import TechnicalFeedback
-from .language import Language
-from .language import LanguageTreeNode
-from .media_models import Document
-from .media_models import DocumentForm
-from .organization import Organization
-from .page import Page
-from .page import PageTranslation
-from .poi import POI
-from .poi import POITranslation
-from .push_notification import PushNotification
-from .push_notification import PushNotificationTranslation
-from .region import Region
-from .user_profile import UserProfile
+from .config.configuration import Configuration
 
-from .user_mfa import UserMfa
+from .events.event import Event
+from .events.event_translation import EventTranslation
+from .events.recurrence_rule import RecurrenceRule
+
+from .extras.extra import Extra
+from .extras.extra_template import ExtraTemplate
+
+from .feedback.event_feedback import EventFeedback
+from .feedback.extra_feedback import ExtraFeedback
+from .feedback.feedback import Feedback
+from .feedback.page_feedback import PageFeedback
+from .feedback.region_feedback import RegionFeedback
+from .feedback.search_result_feedback import SearchResultFeedback
+from .feedback.technical_feedback import TechnicalFeedback
+
+from .languages.language import Language
+from .languages.language_tree_node import LanguageTreeNode
+
+from .media.document import Document
+
+from .pages.page import Page
+from .pages.page_translation import PageTranslation
+
+from .pois.poi import POI
+from .pois.poi_translation import POITranslation
+
+from .push_notifications.push_notification import PushNotification
+from .push_notifications.push_notification_translation import PushNotificationTranslation
+
+from .regions.region import Region
+
+from .users.organization import Organization
+from .users.user_profile import UserProfile
+from .users.user_mfa import UserMfa
