+++ backend/cms/views/users/users.py
@@ -7,9 +7,9 @@ from django.utils.decorators import method_decorator
 from django.utils.translation import ugettext as _
 from django.views.generic import TemplateView
 
-from ...models import UserProfile
-from ...forms.users import UserForm, UserProfileForm
 from ...decorators import staff_required
+from ...forms.users import UserForm, UserProfileForm
+from ...models import UserProfile
 
 
 @method_decorator(login_required, name='dispatch')
